pyproject.toml
## Summary: Project build configuration specifying dependencies and package metadata.
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lifeimage"
version = "0.1.0"
description = "Life Engine optimization to approximate target image"
authors = [{name="LifeImage"}]
dependencies = [
  "numpy>=1.26,<2.0",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "imageio>=2.34"
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
addopts = "-q"
